+++ src/app/g-maps/g-maps.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit, ViewChild, ElementRef } from '@angular/core';
+import { DataService } from '../data/shared/data.service';
 import { GMapService } from './shared/g-map.service';
 import { Polygon } from './shared/polygon';
+import { Region } from './shared/region';
 
 @Component({
   selector: 'app-g-maps',
@@ -11,19 +13,30 @@ export class GMapsComponent implements OnInit {
 
   @ViewChild('gmap') gMapElement: ElementRef;
 
-  constructor(private gMapService: GMapService) { }
+  regions: Region[];
+
+  constructor(private gMapService: GMapService, private dataService: DataService) { }
 
   ngOnInit() {
+    // Initialize Google Maps
     this.gMapService.initialize(new google.maps.Map(this.gMapElement.nativeElement, {
       center: new google.maps.LatLng(43.6006866, 1.4390260), // Toulouse, France
-      zoom: 12.5,
+      zoom: 12,
       mapTypeId: google.maps.MapTypeId.ROADMAP
     }));
-    this.gMapService.addPolygon(new Polygon([
-      {lat: 43.590, lng: 1.439},
-      {lat: 43.600, lng: 1.449},
-      {lat: 43.610, lng: 1.429}
-    ], '#FF0000'));
+    // Load data and display it on the map
+    this.dataService.getRegions().subscribe(data => {
+      this.regions = data;
+      this.regions.forEach(region => this.gMapService.addPolygon(region.polygon, region.code));
+    });
+  }
+
+  filter(min: number, max: number) {
+    min = min || 0;
+    max = max || 100;
+    this.regions.forEach(region => {
+      region.polygon.getGPolygon().setVisible(region.score >= min && region.score <= max);
+    });
   }
 
 }