+++ src/app/g-maps/g-maps.module.ts
@@ -2,10 +2,13 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { GMapsComponent } from './g-maps.component';
 import { GMapService } from './shared/g-map.service';
+import { HttpClientModule } from '@angular/common/http';
+import { DataModule } from '../data/data.module';
 
 @NgModule({
   imports: [
-    CommonModule
+    CommonModule,
+    DataModule
   ],
   providers: [GMapService],
   declarations: [GMapsComponent],