+++ src/app/g-maps/shared/region.ts
@@ -1,0 +1,8 @@
+import { Polygon } from './polygon';
+
+export class Region {
+  id: number;
+  code: string;
+  score: number; // Score out of 100
+  polygon: Polygon;
+}