+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+<a name="0.1.0"></a>
+# [0.1.0](https://github.com/rcsty/data-map/compare/0.0.1...0.1.0) (2018-04-15)
+
+### Features
+
+* **http:** Fetch data from backend (use mocked data if not)
+* **map:** Shows an info window when a polygon is clicked
+* **filter:** Add a filter
+
 <a name="0.0.1"></a>
 # [0.0.1](https://github.com/rcsty/data-map/compare/0.0.0...0.0.1) (2018-04-15)